requirements.txt
click==8.1.3
Flask==2.2.2
flask-sock==0.5.2
h11==0.13.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
pyparsing==3.0.9
pyserial==3.5
python-can==4.0.0
simple-websocket==0.8.1
typing-extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.1
